refactor: NotificationController, NotificationService 코드 수정
